requirements.txt
Django==3.0.5
django_compressor>=2.4
django-libsass==0.8
django-crispy-forms==1.9
pillow==7.1
django-summernote==0.8.11.6
django-crum==0.7.2